Added important privacy warning re: Sawbuck
It seems appropriate to caution users with regard to posting raw Sawbuck output online. It's not immediately obvious that this tool captures output from all running Chrome instances (and not just the Gitter Desktop client).

I've proposed the specific language; feel free to edit as needed (or not to use it at all, in which case, *something* should be stated to this effect).

Thanks!